+++ src/toy.ts
@@ -1,2 +1,0 @@
-const obj = JSON.parse('{"a": 12}');
-console.log(obj.b);