+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.2.0
+
+- Re-architected to use VS Code's built-in TypeScript server (thanks for the tip @acutmore!).
+- Added an icon (thanks ChatGPT!)
+- Removed `anyXray.renderErrorAnys` setting, which no longer makes sense in the new architecture.
+- Switched from TypeScript -> Babel for parsing. This results in a significantly smaller extension.
+
 ## 0.1.1
 
 - Reduce VS Code requirement to support Cursor